Caboo_backend/requirements.txt
Django
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
celery
django-celery-results
django-celery-beat
channels
channels-redis
daphne
python-dotenv
razorpay
google-auth
redis
requests
requests-oauthlib
python-decouple
social-auth-app-django
